Update Program.cs
I have finally managed to take the process out in two methods, however the output is still not correct. Though it is close to what i require.

I believe the next step will be to create out of the two just one method which will take 2 inputs and give two outputs at the same time.